se cargan las bicicletas

+++ src/api/models.py
@@ -54,8 +54,8 @@ class Bike(db.Model):
 class Estaciones(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     direccion = db.Column(db.String(120), unique=True, nullable=False)
-    numeroBicis = db.Column(db.Integer(80))
-    numeroParkings = db.Column(db.Integer(80))
+    numeroBicis = db.Column(db.Integer)
+    numeroParkings = db.Column(db.Integer)
 
 
     def __repr__(self):
